+++ tests/test_glob.py
@@ -623,8 +623,8 @@ class Testglob(_TestGlob):
         Options(absolute=True, skip=sys.platform != "win32"),
         [('*:',), []],
         [('?:',), []],
-        [(r'\\\\?\\c:\\',), [('\\\\?\\c:\\',)]],
-        [(r'\\\\*\\*\\',), []],
+        [(R'\\\\?\\c:\\',), [('\\\\?\\c:\\',)]],
+        [(R'\\\\*\\*\\',), []],
         Options(absolute=False, skip=False),
 
         Options(skip=not can_symlink()),
@@ -659,7 +659,7 @@ class Testglob(_TestGlob):
             ]
         ],
         [
-            (r'*\\',),
+            (R'*\\',),
             [
                 ('aab', ''), ('aaa', ''), ('a', '')
             ] if not can_symlink() else [
@@ -675,7 +675,7 @@ class Testglob(_TestGlob):
         [('[a]',), [('a',)]],
         [('[!b]',), [('a',)]],
         [('[^b]',), [('a',)]],
-        [(r'@([\a]|\aaa)',), [('a',), ('aaa',)]],
+        [(R'@([\a]|\aaa)',), [('a',), ('aaa',)]],
 
         Options(absolute=True),
         # Test empty.
@@ -1014,6 +1014,91 @@ class Testglob(_TestGlob):
             self.assert_equal(sorted(glob.glob(['dummy', '**/aab'], flags=glob.G)), ['aab',])
 
 
+class TestGlobStarLong(_TestGlob):
+    """Test `***` cases."""
+
+    DEFAULT_FLAGS = glob.BRACE | glob.EXTGLOB | glob.GLOBSTARLONG | glob.MARK | glob.NOUNIQUE
+
+    cases = [
+        # Test `globstar` with `globstarlong` enabled
+        [
+            ('**', '*'),
+            [
+                ('aab',), ('aab', 'F'), ('a',), ('a', 'bcd'), ('a', 'bcd', 'EF'), ('a', 'bcd', 'efg'),
+                ('a', 'bcd', 'efg', 'ha'), ('a', 'D'), ('aaa',), ('aaa', 'zzzF'), ('EF',), ('ZZZ',)
+            ] if not can_symlink() else [
+                ('aab',), ('aab', 'F'), ('a',), ('a', 'bcd'), ('a', 'bcd', 'EF'), ('a', 'bcd', 'efg'),
+                ('a', 'bcd', 'efg', 'ha'), ('a', 'D'), ('aaa',), ('aaa', 'zzzF'), ('EF',), ('ZZZ',),
+                ('sym1',), ('sym2',), ('sym3',)
+            ]
+        ],
+
+        # Test `globstarlong`
+        [
+            ('***', '*'),
+            [
+                ('aab',), ('aab', 'F'), ('a',), ('a', 'bcd'), ('a', 'bcd', 'EF'), ('a', 'bcd', 'efg'),
+                ('a', 'bcd', 'efg', 'ha'), ('a', 'D'), ('aaa',), ('aaa', 'zzzF'), ('EF',), ('ZZZ',)
+            ] if not can_symlink() else [
+                ('aab',), ('aab', 'F'), ('a',), ('a', 'bcd'), ('a', 'bcd', 'EF'), ('a', 'bcd', 'efg'),
+                ('a', 'bcd', 'efg', 'ha'), ('a', 'D'), ('aaa',), ('aaa', 'zzzF'), ('EF',), ('ZZZ',),
+                ('sym1',), ('sym2',), ('sym3',), ('sym3', 'EF'), ('sym3', 'efg'), ('sym3', 'efg', 'ha')
+            ]
+        ],
+
+        # Enable `FOLLOW`. Should be no changes.
+        [
+            ('**', '*'),
+            [
+                ('aab',), ('aab', 'F'), ('a',), ('a', 'bcd'), ('a', 'bcd', 'EF'), ('a', 'bcd', 'efg'),
+                ('a', 'bcd', 'efg', 'ha'), ('a', 'D'), ('aaa',), ('aaa', 'zzzF'), ('EF',), ('ZZZ',)
+            ] if not can_symlink() else [
+                ('aab',), ('aab', 'F'), ('a',), ('a', 'bcd'), ('a', 'bcd', 'EF'), ('a', 'bcd', 'efg'),
+                ('a', 'bcd', 'efg', 'ha'), ('a', 'D'), ('aaa',), ('aaa', 'zzzF'), ('EF',), ('ZZZ',),
+                ('sym1',), ('sym2',), ('sym3',)
+            ],
+            glob.L
+        ],
+        [
+            ('***', '*'),
+            [
+                ('aab',), ('aab', 'F'), ('a',), ('a', 'bcd'), ('a', 'bcd', 'EF'), ('a', 'bcd', 'efg'),
+                ('a', 'bcd', 'efg', 'ha'), ('a', 'D'), ('aaa',), ('aaa', 'zzzF'), ('EF',), ('ZZZ',)
+            ] if not can_symlink() else [
+                ('aab',), ('aab', 'F'), ('a',), ('a', 'bcd'), ('a', 'bcd', 'EF'), ('a', 'bcd', 'efg'),
+                ('a', 'bcd', 'efg', 'ha'), ('a', 'D'), ('aaa',), ('aaa', 'zzzF'), ('EF',), ('ZZZ',),
+                ('sym1',), ('sym2',), ('sym3',), ('sym3', 'EF'), ('sym3', 'efg'), ('sym3', 'efg', 'ha')
+            ],
+            glob.L
+        ]
+    ]
+
+    @classmethod
+    def setup_fs(cls):
+        """Setup file system."""
+
+        cls.mktemp('a', 'D')
+        cls.mktemp('aab', 'F')
+        cls.mktemp('.aa', 'G')
+        cls.mktemp('.bb', 'H')
+        cls.mktemp('aaa', 'zzzF')
+        cls.mktemp('ZZZ')
+        cls.mktemp('EF')
+        cls.mktemp('a', 'bcd', 'EF')
+        cls.mktemp('a', 'bcd', 'efg', 'ha')
+        cls.can_symlink = can_symlink()
+        if cls.can_symlink:
+            os.symlink(cls.norm('broken'), cls.norm('sym1'))
+            os.symlink('broken', cls.norm('sym2'))
+            os.symlink(os.path.join('a', 'bcd'), cls.norm('sym3'))
+
+    @pytest.mark.parametrize("case", cases)
+    def test_glob_cases(self, case):
+        """Test glob cases."""
+
+        self.eval_glob_cases(case)
+
+
 class TestGlobMarked(Testglob):
     """Test glob marked."""
 
@@ -1335,14 +1420,14 @@ class TestGlobCornerCase(_TestGlob):
         [('@(a/b)',), []],
         [('@(a[/]b)',), []],
         [('test[',), [('test[',)]],
-        [(r'a\/b',), [('a', 'b')]],
-        [(r'a[\/]b',), [('a[', ']b')]],
+        [(R'a\/b',), [('a', 'b')]],
+        [(R'a[\/]b',), [('a[', ']b')]],
 
         Options(skip=util.is_case_sensitive()),
         [('a[\\',), [('a[',)]],
         [('@(a[\\',), [('@(a[',)]],
-        [(r'a[\\',), [('a[', '')]],
-        [(r'@(a[\\',), [('@(a[', '')]],
+        [(R'a[\\',), [('a[', '')]],
+        [(R'@(a[\\',), [('@(a[', '')]],
         Options(skip=False)
     ]
 
@@ -1397,53 +1482,53 @@ class TestGlobEscapes(unittest.TestCase):
 
         check = self.check_escape
         check('abc', 'abc')
-        check('[', r'\[')
-        check('?', r'\?')
-        check('*', r'\*')
-        check('[[_/*?*/_]]', r'\[\[_/\*\?\*/_\]\]')
-        check('/[[_/*?*/_]]/', r'/\[\[_/\*\?\*/_\]\]/')
+        check('[', R'\[')
+        check('?', R'\?')
+        check('*', R'\*')
+        check('[[_/*?*/_]]', R'\[\[_/\*\?\*/_\]\]')
+        check('/[[_/*?*/_]]/', R'/\[\[_/\*\?\*/_\]\]/')
 
     @unittest.skipUnless(sys.platform.startswith('win'), "Windows specific test")
     def test_escape_windows(self):
         """Test windows escapes."""
 
         check = self.check_escape
-        check('a:\\?', r'a:\\\?')
-        check('b:\\*', r'b:\\\*')
-        check('\\\\?\\c:\\?', r'\\\\?\\c:\\\?')
-        check('\\\\*\\*\\*', r'\\\\*\\*\\\*')
-        check('//?/c:/?', r'//?/c:/\?')
-        check('//*/*/*', r'//*/*/\*')
-        check('//[^what]/name/temp', r'//[^what]/name/temp')
+        check('a:\\?', R'a:\\\?')
+        check('b:\\*', R'b:\\\*')
+        check('\\\\?\\c:\\?', R'\\\\?\\c:\\\?')
+        check('\\\\*\\*\\*', R'\\\\*\\*\\\*')
+        check('//?/c:/?', R'//?/c:/\?')
+        check('//*/*/*', R'//*/*/\*')
+        check('//[^what]/name/temp', R'//[^what]/name/temp')
 
     def test_escape_forced_windows(self):
         """Test forced windows escapes."""
 
         check = self.check_escape
-        check('a:\\?', r'a:\\\?', unix=False)
-        check('b:\\*', r'b:\\\*', unix=False)
-        check('\\\\?\\c:\\?', r'\\\\?\\c:\\\?', unix=False)
-        check('\\\\*\\*\\*', r'\\\\*\\*\\\*', unix=False)
-        check('//?/c:/?', r'//?/c:/\?', unix=False)
-        check('//*/*/*', r'//*/*/\*', unix=False)
+        check('a:\\?', R'a:\\\?', unix=False)
+        check('b:\\*', R'b:\\\*', unix=False)
+        check('\\\\?\\c:\\?', R'\\\\?\\c:\\\?', unix=False)
+        check('\\\\*\\*\\*', R'\\\\*\\*\\\*', unix=False)
+        check('//?/c:/?', R'//?/c:/\?', unix=False)
+        check('//*/*/*', R'//*/*/\*', unix=False)
         check(
             '//./Volume{b75e2c83-0000-0000-0000-602f00000000}/temp',
-            r'//./Volume\{b75e2c83-0000-0000-0000-602f00000000\}/temp',
+            R'//./Volume\{b75e2c83-0000-0000-0000-602f00000000\}/temp',
             unix=False
         )
-        check('//[^what]/name/temp', r'//[^what]/name/temp', unix=False)
+        check('//[^what]/name/temp', R'//[^what]/name/temp', unix=False)
 
     def test_escape_forced_unix(self):
         """Test forced windows Unix."""
 
         check = self.check_escape
-        check('a:\\?', r'a:\\\?', unix=True)
-        check('b:\\*', r'b:\\\*', unix=True)
-        check('\\\\?\\c:\\?', r'\\\\\?\\c:\\\?', unix=True)
-        check('\\\\*\\*\\*', r'\\\\\*\\\*\\\*', unix=True)
-        check('//?/c:/?', r'//\?/c:/\?', unix=True)
-        check('//*/*/*', r'//\*/\*/\*', unix=True)
-        check('//[^what]/name/temp', r'//\[^what\]/name/temp', unix=True)
+        check('a:\\?', R'a:\\\?', unix=True)
+        check('b:\\*', R'b:\\\*', unix=True)
+        check('\\\\?\\c:\\?', R'\\\\\?\\c:\\\?', unix=True)
+        check('\\\\*\\*\\*', R'\\\\\*\\\*\\\*', unix=True)
+        check('//?/c:/?', R'//\?/c:/\?', unix=True)
+        check('//*/*/*', R'//\*/\*/\*', unix=True)
+        check('//[^what]/name/temp', R'//\[^what\]/name/temp', unix=True)
 
 
 @unittest.skipUnless(sys.platform.startswith('win'), "Windows specific test")
@@ -1573,7 +1658,7 @@ class TestGlobPaths(unittest.TestCase):
         self.assertTrue(len(results) > 0)
         self.assertTrue('\\' not in results)
 
-        results = glob.glob(r'\\*')
+        results = glob.glob(R'\\*')
         self.assertTrue(len(results) > 0)
         self.assertTrue('\\' not in results)