+++ docs/src/markdown/about/changelog.md
@@ -2,9 +2,13 @@
 
 ## 10.0
 
+-   **NEW**: Added `GLOBSTARLONG` which adds support for the Zsh style `***` which acts like `**` with `GLOBSTAR` but
+    but traverses symlinks.
+-   **NEW**: `pathlib.match` will respect symlinks (when the `REALPATH` flag is given) and hidden file rules (enable
+    `DOTALL` to match hidden files).
 -   **NEW**: Symlinks should not be traversed when `GLOBSTAR` is enabled unless `FOLLOW` is also enabled, but they
     should still be matched. Prior to this change, symlinks were not traversed _and_ they were ignored from matching
-    which contradicts how Bash works, which is are general target.
+    which contradicts how Bash works and could be confusing to users.
 -   **FIX**: Fix some inconsistencies with `globmatch` and symlink handling when `REALPATH` is enabled.
 
 ## 9.0